requirements.txt
aiogram>=3.13.1
openai>=1.51.0
aiosqlite>=0.20.0
python-dotenv>=1.0.1
cryptography>=43.0.1
Pillow>=10.4.0
